foundation-skeleton/backend/pyproject.toml
[project]
name = "mcp-backend"
version = "0.1.0"
description = "MCP Backend API"
requires-python = ">=3.11"

[project.dependencies]
fastapi = "0.115.12"
uvicorn = {extras = ["standard"], version = "0.34.0"}
python-jose = "3.3.0"
python-multipart = "0.0.20"
redis = "5.2.1"
sqlalchemy = {extras = ["asyncio"], version = "2.0.39"}
pydantic = {extras = ["email"], version = "2.10.6"}
pydantic-settings = "2.8.1"
alembic = "1.15.2"
psycopg = {extras = ["binary"], version = "3.2.6"}
email-validator = "2.2.0"
opentelemetry-api = "1.31.1"
opentelemetry-sdk = "1.31.1"
opentelemetry-exporter-otlp = "1.31.1"
opentelemetry-instrumentation-fastapi = "0.52b1"
opentelemetry-instrumentation-logging = "0.52b1"
opentelemetry-instrumentation-redis = "0.52b1"
opentelemetry-instrumentation-sqlalchemy = "0.52b1"

[project.optional-dependencies]
test = [
    "pytest==8.3.5",
    "pytest-asyncio==0.24.0",
    "aiosqlite==0.21.0",
    "httpx==0.28.1"
]
dev = ["ipython"]
telemetry = [
    "opentelemetry-api==1.31.1",
    "opentelemetry-sdk==1.31.1",
    "opentelemetry-exporter-otlp==1.31.1",
    "opentelemetry-instrumentation-fastapi==0.52b1",
    "opentelemetry-instrumentation-logging==0.52b1",
    "opentelemetry-instrumentation-redis==0.52b1",
    "opentelemetry-instrumentation-sqlalchemy==0.52b1"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
